add comments/stubs for using fira or STIX2 (sans serif) fonts.
Sans serif fonrs are more dense and they should also be a bit easier to read
on something like a cheatsheet.

But as I'm not completely sure what I prefere, I only added this as
comments.

Note: Fira is (sadly) not as extensive as STIX2 when defining symbols.
Therefore some symbols might not be available and you need to manually
define them with another font (e.g. STIX2) if you want to use these symbols.